+++ src/components/nuclei/sperm/RodentSpermNucleus.java
@@ -68,11 +68,11 @@ public class RodentSpermNucleus extends SpermNucleus {
 		super();
 	}
 	
-	public RodentSpermNucleus (Roi roi, File file, int number, double[] position) { // construct from an roi
+	public RodentSpermNucleus (Roi roi, File file, int number, int[] position) { // construct from an roi
 		super(roi, file, number, position);
 	}
 	
-	public RodentSpermNucleus (Roi roi, File file, int number, double[] position, IPoint centreOfMass) { // construct from an roi
+	public RodentSpermNucleus (Roi roi, File file, int number, int[] position, IPoint centreOfMass) { // construct from an roi
 		super(roi, file, number, position, centreOfMass);
 	}
